+++ app/src/component/Layout/Header.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Menu, Layout } from "antd"
 import SubMenu from "antd/lib/menu/SubMenu"
 import { logout } from "~/ApiServices/Login"
-import styles from "~/component/Offering/Header.module.scss"
+import styles from "~/component/Layout/Header.module.scss"
 
 const { Header: AntdHeader } = Layout
