+++ app/src/component/Offering/SelectedFilters.module.scss
