+++ app/src/component/Offering/index.ts
@@ -1,4 +1,2 @@
-export * from "~/component/Offering/Header"
-export * from "~/component/Offering/BreadCrumb"
-export * from "~/component/Offering/Toolbar"
+export * from "~/component/Offering/SelectedFilters"
 export * from "~/component/Offering/FilterColumn"
